change name folder training_sequential => training
