+++ nachosv2/training/training/config_files/mouse_data/mouse_data_1.yml
