+++ nachosv2/training/training/config_files/mouse_data/single_configuration_1.yml
