--- nachosv2/training/training_multiprocessing/mpi_config.json
@@ -1,6 +1,0 @@
-{
-    "cuda_devices": [0, 1],
-    "gpu_addrs": [],
-    "n_processes": 2,
-    "is_outer": true
-}
