--- nachosv2/training/training_sequential/config_files/2D_config_inner_InceptionV3_trial_parallel.json
@@ -1,37 +1,0 @@
-{
-    "hyperparameters": {
-        "batch_size": 32,
-        "channels": 1,
-        "cropping_position": [40, 10],
-        "decay": 0.01,
-        "do_cropping": false,
-        "epochs": 20,
-        "learning_rate": 0.01,
-        "momentum": 0.9,
-	    "bool_nesterov": true,
-        "patience": 10
-    },
-
-    "data_input_directory": "/home/pcallec/mif_outer/data/pig_kidney_subset",
-    "csv_input_directory": "data/2D_kidney_csv",
-    "output_path": "results/distributed/pig_kidney_subset_parallel",
-    "job_name": "pig_kidney_subset",
-    
-    "k_epoch_checkpoint_frequency": 1,
-
-    "shuffle_the_images": true,
-    "shuffle_the_folds": false,
-    "seed": 9,
-
-    "class_names": ["cortex", "medulla", "pelvis-calyx"],
-    "selected_model_name": "InceptionV3",
-    "metrics": [],
-    
-    "subject_list": ["k1", "k2", "k3", "k4", "k5"],
-    "test_subjects": ["k1", "k2", "k3", "k4", "k5"],
-    "validation_subjects": ["k1", "k2", "k3", "k4", "k5"],
-    
-    "image_size": [301, 235],
-    "target_height": 301,
-    "target_width": 235
-}
