--- nachosv2/training/training_sequential/config_files/OCT_small/oct_small_3.json
@@ -1,40 +1,0 @@
-{
-    "hyperparameters": {
-        "batch_size": 32,
-        "number_channels": 1,
-        "cropping_position": [40, 10],
-        "decay": 0.01,
-        "do_cropping": false,
-        "epochs": 100,
-        "learning_rate": 0.01,
-        "momentum": 0.9,
-	    "bool_nesterov": true,
-        "patience": 20
-    },
-
-    "data_input_directory": "/home/pcallec/mif_outer/data/pig_kidney_subset",
-    "path_metadata_csv": "/home/pcallec/NACHOS/data/2D_kidney_csv/all.csv",
-    "output_path": "/home/pcallec/NACHOS/results/pig_kidney_subset_all_folds_v2",
-    "job_name": "pig_kidney_subset",
-    
-    "checkpoint_epoch_frequency": 20,
-
-    "do_normalize_2d": true,
-    "shuffle_the_images": true,
-    "shuffle_the_folds": false,
-    "seed": 9,
-
-    "use_mixed_precision": false,
-
-    "class_names": ["cortex", "medulla", "pelvis-calyx"],
-    "architecture_name": "InceptionV3",
-    "metrics_list": [],
-    
-    "fold_list": ["k1", "k2", "k3", "k4", "k5"],
-    "test_fold_list": ["k1", "k2", "k3", "k4", "k5"],
-    "validation_fold_list": ["k1", "k2", "k3", "k4", "k5"],
-    
-    "image_size": [301, 235],
-    "target_height": 301,
-    "target_width": 235
-}
